add packaging information to version
Summary:
Add a "source version" when generating source packages for Ubuntu. The source
version is a part of the software version, which is encoded in the original
tarball. This allows some flexibility when uploading to Launchpad, which will
reject changes for the same software version. For example, the source version
can be bumped to upload a fix for a broken package (without bumping the actual
software version). Note that the package version cannot be used to accomplish
this; it is not part of the source tarball, so bumping this version does not
pervent Launchpad from rejecting the changes.

Rename the script, since this is a bit specific for Ubuntu and Launchpad.

See: http://askubuntu.com/a/391751

Reviewers: jacko

Reviewed By: jacko

Differential Revision: https://phabricator.buildinspace.com/D163