add build deprecation warning
Summary: Doing this before we actually remove the build field is kinder to users.

Test Plan:
This dumps a bunch of junk into our test output, because we still use
the `build` field. I don't think it's worth putting in a bunch of extra
logic to catch that. This won't be around for very long, and this will
be an incentive to clean it up.

Reviewers: sean

Reviewed By: sean

Differential Revision: https://phabricator.buildinspace.com/D165