simplify Debian versioning
Summary:
The package version was being appended to the version given to `dch`. This
complicated versioning for no good reason. This change collapses the version of
peru, series, and a package version into one unified version that is used with
all the tools (including `dch`).

"There can be only one!"

Reviewers: jacko

Reviewed By: jacko

Differential Revision: https://phabricator.buildinspace.com/D166