create scope.py
Summary:
The new Scope class takes over target parsing from `imports.py`, and it
takes over holding the dictionaries of defined modules and rules from
`Runtime`. The target parsing methods it provides are capable of
splitting up a scoped name (i.e. `a.b.c`) and traversing the containing
modules. This depends on a `Module.get_scope` method that doesn't exist
yet, but we mock it out for the test.

This is the last major piece of groundwork we need to lay before we can
start enabling recursive features.

Right now we stick `imports` onto `Scope`, just because it's parsed at
the same time as `modules` and `rules`. That doesn't feel quite right,
and I'm planning to clean it up in subsequent diffs.

Test Plan:
A new unit test that mocks out modules, rules, and scope creation. We
only directly test `parse_target`, but that depends on most of the rest
of the functionality in `Scope`.

Reviewers: sean

Reviewed By: sean

Differential Revision: https://phabricator.buildinspace.com/D174