implement Module.get_scope and enable scoped imports!
Summary:
This is the real version of what we mocked out in the previous diff. Use
the `Cache.read_file` method that's been waiting around forever to read
the `peru.yaml` file from a fetched module, and then parse that to get
that module's scope. With this change, now you can import a target like
`a.b.c`.

There was a bug in `parser.py` where we tried to validate a name *after*
applying the scope prefix, which caused an error because names can't
contain `.`. Moving the check earlier fixes that.

Test Plan: New integration test.

Reviewers: sean

Reviewed By: sean

Differential Revision: https://phabricator.buildinspace.com/D175