fill out Module.get_tree and enable recursive imports!
Summary:
Implement the `Module.get_tree` method that we turned into a stub in the
previous diff. This uses the same `peru.yaml` parsing code and assembles
a full imports tree out of it, which it merges with the base tree to
produce the full tree of a peru project. Factor out the
`get_imports_tree` function to help with this.

Test Plan: New integration test.

Reviewers: sean

Reviewed By: sean

Differential Revision: https://phabricator.buildinspace.com/D176