add `pick` field
Summary:
Add a `pick` field that allows files to be selected from the module
tree. Note that paths are relative to the module root, NOT the export path,
unlike the deprecated `files` field. This means that all paths are "absolute"
and applied before export.

Test Plan: Carbon copy of `files` tests.

Reviewers: jacko

Reviewed By: jacko

Differential Revision: https://phabricator.buildinspace.com/D180