add the `move` field
Summary:
This field is compensating for some of the functionality we lose by
getting rid of `build`. See https://github.com/buildinspace/peru/issues/83.

There are security issues with the code as it stands. We need to prevent
destinations from being absolute paths, or symlinks that point out of
the project. See https://github.com/buildinspace/peru/issues/87. I plan
to address that for a bunch of fields together in a subsequent diff.

Test Plan: New integration test.

Reviewers: sean

Reviewed By: sean

Differential Revision: https://phabricator.buildinspace.com/D185