make move and copy into tree operations
Summary:
These are the first callers of the new `ls_tree` and `modify_tree`
methods. To preserve the familiar `mv`/`cp` behavior, we check whether
destinations are directories, and if so we put files into those
directories rather than overwriting.

I've gone ahead and allowed `move` to take an optional multimap, so that
we can delete some code in the parser. This is conceptually a little
strange, but it was already possible to move both a directory to one
place and its contents to another and end up with copies (because of the
all-at-once way that moves and copies are implemented). So keeping in
extra code to try to prevent multi-moves didn't seem worth it.

Test Plan: Create `test_rules.py` and add a couple tests.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D200