create glob.py
Summary:
For operations on git trees, we'll need a way to expand globs that
doesn't touch the filesystem. Having played with git's native globbing
before, it's not great. (I don't think ** is allowed to be empty, for
example.) This creates a new `peru.glob` module, which can convert path
globs into regexes.

Test Plan: Tons of test cases.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D190