do more explicit typechecking in the parser
Summary:
Some typos in `peru.yaml` cause the parser to assume something is a
dictionary when in fact it's just a string or something. The errors that
result are really difficult for users to understand. To catch at least
some of these cases, always check a dictionary's type before calling
`pop`.

Closes https://github.com/buildinspace/peru/issues/96.

Test Plan: One new tests to at least check the case in the original issue.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D204