stop capturing stderr in plugins
Normally plugins allow all subprocess output to go to the caller. Peru
puts all that output in the fancy display (or buffers it and prints it
later, in --verbose mode). Sometimes plugins need to read the output of
a command, though, like `git rev-parse`. In those cases, previously we
were capturing both stdout and stderr. But we shouldn't do that --
whatever ends up on stderr is likely to screw up the parsing that the
plugins was intending to do, and there's no reason to prevent stderr
messages from reaching peru's display/log. With this diff, plugins only
ever capture stdout.

Note that there's a bit of repeated code in the git/hg/svn plugins. It's
not too bad, just code for shelling out, but it does mean that we need
to make this fix in three places. Something to keep in mind.

Closes https://github.com/buildinspace/peru/issues/105.