add a warning for duplicate keys
PyYAML allows duplicate keys to silently overwrite each other. This is
almost never what you want, and it can be confusing for new users. It
would be nice to have a YAML parser that throws in this case, but until
then we can use a little heuristic to try to detect this and print a
warning.

Go line by line in the peru.yaml file, and keep track of what keys have
been seen so far at the current indentation level. If we see the same
one twice, print the warning. This could get confused by multiline
strings, JSON-style dicts, and who knows what else, but it's just a
warning. Suppress it with --quiet.

Closes https://github.com/buildinspace/peru/issues/107.