fix executable flag tests on Windows
Previously we were checking for Windows in tests that looked for the
executable flag (which doesn't exist there), and I forgot one. Create a
couple helper assertions that check these flags and automatically handle
Windows, and use those instead of forcing every test to remember to do
the right thing.