implement --peru-file, --sync-dir, etc.
Summary:
We've been meaning to do this for a while, see
https://github.com/buildinspace/peru/issues/108. Here's the full set:
PERU_FILE      --peru-file
PERU_SYNC_DIR  --sync-dir
PERU_STATE_DIR --state-dir
PERU_CACHE_DIR --cache-dir

Note that --peru-file and --sync-dir need to be set together, or not at
all; you can't just specify one of them. That's because if you were to
set the one that's normally derived from the other (which we could
imagine defining in either direction), you end up with weird
consequences for the first one that's still derived from the cwd.

Test Plan:
Several new integration tests. In the process, I found out that we had
a few cases where we passed a string instead of a tuple to test
helper methods. Added a new assertion to catch those.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D210