split the docopt parser into subcommands
Summary:
This lets us output command-specific help, which is very nice. It also
is a bit of a prerequisite for adding the big set of flags like
`--sync-dir`, because we really didn't want to add all N of those to
every subcommand independently.

Test Plan:
Added some new tests for the help output, in particular that
`peru -h sync` and `peru sync -h` do the same thing.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D209