handle comments in our approximate-YAML-parsing code
We kinda-sorta parse YAML files ourselves to look for duplicate keys.
This code isn't robust at all, but it's only used to generate a warning,
so it's probably better than writing a full-blown YAML parser.

One mistake this parser was making was treating commented rows as
top-level keys, and then wrongly reporting duplicates. This is a quick
workaround for that case.