always update the index file on the slow path
There's a case I'd missed when I wrote the current fast-path-optimized
version of Cache.export(). If a file has been touched, but not actually
changed, it will show up as modified in the first call to diff-files.
That kicks us down the slow path. However, the slow path doesn't
actually call update-index, so the second call to diff-files (which
filters out deletions) *also* shows it as modified, which makes peru
think the working copy is dirty and error out.

Fix this issue by guaranteeing that the slow path will call
update-index, if it wasn't called before.