disable recursion by default, and update the warning
It turns out that recursion-by-default interacts poorly with one of
peru's core design goals, that it should be easy to call from other
cools. Consider this example:

1. Project A has some peru dependencies and a Makefile, so it runs
   `peru sync` inside the Makefile.
2. Project B has the same structure, and also B depends A. That means
   that B's peru.yaml file imports A, and also B's Makefile calls A's
   Makefile.
3. You clone Project B and run `make`.

The first thing that will happen is probably B's Makefile calling `peru
sync`. Now if recursion is on, that sync will *also* fetch A's files.
The next thing that happens might be B's Makefile calling into A's
Makefile, which is going to try to `peru sync` again. This is the
problem. A's dependencies are already in place, because of the recursive
sync, and when A tries to sync them again it's going to run into an
existing files conflict.

The core issue is that when peru is wrapped by other tools, it's not ok
to reach through those abstractions and aggressively sync stuff. We
could maybe come up with some fancy logic to detect when an outer
project has already synced us, to avoid the conflicts in this case, but
there are more complicated situations where that doesn't help. What if
the inner project decides to sync something *different*? What state
would that put the outer project in?

Instead of confronting this chaos, we're disabling recursive fetching by
default. It's still available to projects that want it, but it won't
screw up projects that don't expect it.

This is a breaking change, and we've introducing a warning about it in
c10c368. The next release that includes this commit should be 0.4.0.
We'll probably remove the warning in 0.5 or 1.0, whichever comes first.

See also https://github.com/buildinspace/peru/issues/146.