test_delete_and_pick_is_an_error
Dropping everything in a tree was causing modify_tree() to return None,
which lead to a different error when we tried to pass that None back to
git as though it was a valid tree. In retrospect, representing the empty
tree as None was not smart on my part. This diff fixes the modify_tree()
case, though others remain...