fix: includes a verification for the default branch error (#213)
Includes in the reup function a verification for errors when asking for
updates in the default branch. The definition of a default branch is
difficult, and sometimes undefinable. So this function uses a list of
common default branches names and check if they exist in the list. It
returns then this value to the calling function.